added new functions

+++ src/todo_api/views.py
@@ -1,13 +1,14 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status
 from rest_framework.viewsets import ViewSet
 from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework.permissions import AllowAny, IsAuthenticated, IsAuthenticatedOrReadOnly, IsAdminUser
 from django.shortcuts import render, get_object_or_404
-from .serializers import TodoSerializer
-from .models import Todo
+from .serializers import TodoSerializer, MovieSerializer
+from .models import Todo, Movie
 from .permissions import IsOwner
+from rest_framework.generics import GenericAPIView, ListAPIView, CreateAPIView
+from rest_framework import mixins, status
 
 #  rest_framework.views --- APIView
 
@@ -16,8 +17,8 @@ class TodoListApiView(APIView):
     permission_classes = [IsAuthenticated]
 
     def get(self, request, *args, **kwargs):
-        # todos = Todo.objects.filter(user=request.user)
-        todos = Todo.objects.all()
+        todos = Todo.objects.filter(user=request.user)
+        # todos = Todo.objects.all()
         serializer = TodoSerializer(todos, many=True)
         return Response(serializer.data)
 
@@ -25,6 +26,7 @@ class TodoListApiView(APIView):
         
         serializer = TodoSerializer(data=request.data)
         if serializer.is_valid():
+            serializer.validated_data['user'] = request.user
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -36,8 +38,11 @@ class TodoDetailApiView(APIView):
     def get_object(self, todo_id):
         try:
             todo_instance = Todo.objects.get(id=todo_id)
+            if todo_instance.user != self.request.user:
+                todo_instance = None
         except Todo.DoesNotExist:
             todo_instance = None
+
         return todo_instance
 
     def get(self, request, todo_id, *args, **kwargs):
@@ -54,6 +59,10 @@ class TodoDetailApiView(APIView):
             return Response({"error":"todo id doesnt find"}, status=status.HTTP_400_BAD_REQUEST)  
         
         serializer = TodoSerializer(instance=todo_instance, data=request.data)
+
+        if todo_instance.user != request.user:
+            return Response({"error": "Unauthorized access"}, status=status.HTTP_403_FORBIDDEN)
+        
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
@@ -62,7 +71,11 @@ class TodoDetailApiView(APIView):
     def delete(self, request, todo_id, *args, **kwargs):
         todo_instance = self.get_object(todo_id)
         if not todo_instance:
-            return Response({"error":"todo id doesnt find"}, status=status.HTTP_400_BAD_REQUEST)  
+            return Response({"error":"todo id doesnt find"}, status=status.HTTP_400_BAD_REQUEST) 
+
+        if todo_instance.user != request.user:
+            return Response({"error": "Unauthorized access"}, status=status.HTTP_403_FORBIDDEN)
+         
         todo_instance.delete()
         return Response({"message":"To do deleted"})
     
@@ -102,3 +115,73 @@ class TodoViewSet(ViewSet):
         item_instance.delete()
         return Response({"message":"To do deleted"})
         
+
+# GenericApi view
+class RetrieveDeleteItem(GenericAPIView):
+    serializer_class = TodoSerializer
+    queryset = Todo.objects.all()
+
+    def get(self, request, *args, **kwargs):
+        instance = self.get_object()
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)
+
+    def delete(self, request, *args, **kwargs):
+        instance = self.get_object()
+        instance.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    
+
+class CreateListItem(mixins.ListModelMixin, mixins.CreateModelMixin, GenericAPIView):
+    serializer_class = TodoSerializer
+    queryset = Todo.objects.all()
+
+    def get(self, request, *args, **kwargs):
+        return self.list(request, *args, **kwargs)
+    
+    def post(self, request, *args, **kwargs):
+        return self.create(request, *args, **kwargs)
+    
+    
+class RetrieveUpdateDeleteItem(mixins.RetrieveModelMixin,
+                                mixins.UpdateModelMixin, 
+                                mixins.DestroyModelMixin,
+                                GenericAPIView):
+    serializer_class = TodoSerializer
+    queryset = Todo.objects.all()
+
+    def get(self, request, *args, **kwargs):
+        return self.retrieve(request, *args, **kwargs)
+    
+    def put(self, request, *args, **kwargs):
+        return self.update(request, *args, **kwargs)
+    
+    def delete(self, request, *args, **kwargs):
+        return self.destroy(request, *args, **kwargs)
+    
+
+#Custome mixins (uzimizga moslab yozib oladigan mixinslar)
+class MultipleFieldLookupMixin:
+
+    def retrieve(self,request, *args, **kwargs):
+        instance = Todo.objects.filter(user=request.user)
+        serializer = TodoSerializer(instance) # darsdagi kodni to'g'irladik darsda (self.get_serializer(insrtance) yozilib ketgan
+        return Response(serializer.data)
+
+
+
+#Concreate views
+class ListAPIItem(ListAPIView):
+    serializer_class = TodoSerializer
+    queryset = Todo.objects.all()
+
+
+
+class MovieCreateItem(CreateAPIView):
+    serializer_class = MovieSerializer
+    queryset = Movie.objects.all()
+
+
+class MovieListItem(ListAPIView):
+    serializer_class = MovieSerializer
+    queryset = Movie.objects.all()
